+++ fv3net/diagnostics/sklearn_model_performance/__main__.py
@@ -1,10 +1,11 @@
 import argparse
 import os
-import shutil
+import tempfile
 import xarray as xr
-from vcm.cloud.fsspec import get_fs, get_protocol
+import yaml
+
+from vcm.cloud.fsspec import get_fs
 from vcm.cloud.gsutil import copy
-from vcm.cubedsphere.constants import INIT_TIME_DIM
 
 from ..create_report import create_report
 from ..data import merge_comparison_datasets
@@ -17,20 +18,6 @@ from .diagnostics import plot_diagnostics
 from .create_metrics import create_metrics_dataset
 from .plot_metrics import plot_metrics
 
-DATA_VARS = [
-    "dQ1",
-    "dQ2",
-    "sphum",
-    "T",
-    "tsea",
-    "net_precipitation",
-    "net_heating",
-    "net_precipitation_physics",
-    "net_heating_physics",
-    "net_precipitation_ml",
-    "net_heating_ml",
-    "delp",
-]
 DATASET_NAME_PREDICTION = "prediction"
 DATASET_NAME_FV3_TARGET = "C48_target"
 DATASET_NAME_SHIELD_HIRES = "coarsened_high_res"
@@ -43,6 +30,108 @@ DPI_FIGURES = {
     "map_plot_3col": 120,
     "map_plot_single": 100,
 }
+
+
+def _is_remote(path):
+    return path.startswith("gs://")
+
+
+def compute_metrics_and_plot(ds, output_dir, names):
+    # TODO refactor this "and" function into two routines
+    ds_pred = ds.sel(dataset=DATASET_NAME_PREDICTION)
+    ds_test = ds.sel(dataset=DATASET_NAME_FV3_TARGET)
+    ds_hires = ds.sel(dataset=DATASET_NAME_SHIELD_HIRES)
+
+    ds_metrics = create_metrics_dataset(ds_pred, ds_test, ds_hires, names)
+    ds_metrics.to_netcdf(os.path.join(output_dir, "metrics.nc"))
+
+    # TODO This should be another script
+    metrics_plot_sections = plot_metrics(ds_metrics, output_dir, DPI_FIGURES, names)
+
+    diag_report_sections = plot_diagnostics(
+        ds_pred,
+        ds_test,
+        ds_hires,
+        output_dir=output_dir,
+        dpi_figures=DPI_FIGURES,
+        names=names,
+    )
+
+    combined_report_sections = {**metrics_plot_sections, **diag_report_sections}
+    create_report(combined_report_sections, "ml_offline_diagnostics", output_dir)
+
+
+def load_data_and_predict_with_ml(
+    test_data_path,
+    model_path,
+    high_res_data_path,
+    num_test_zarrs,
+    model_type,
+    downsample_time_factor,
+    names,
+):
+
+    # if output path is remote GCS location, save results to local output dir first
+    # TODO I bet this output preparation could be cleaned up.
+    # TODO this function mixes I/O and computation
+    # Should just be 1. load_data, 2. make a prediction
+    ds_test, ds_pred = predict_on_test_data(
+        test_data_path,
+        model_path,
+        num_test_zarrs,
+        names["pred_vars_to_keep"],
+        names["init_time_dim"],
+        names["coord_z_center"],
+        model_type,
+        downsample_time_factor,
+    )
+
+    fs_input = get_fs(args.test_data_path)
+
+    ds_test = add_column_heating_moistening(
+        ds_test,
+        names["suffix_coarse_train_diag"],
+        names["var_pressure_thickness"],
+        names["var_q_moistening_ml"],
+        names["var_q_heating_ml"],
+        names["coord_z_center"],
+    )
+
+    ds_pred = add_column_heating_moistening(
+        ds_pred,
+        names["suffix_coarse_train_diag"],
+        names["var_pressure_thickness"],
+        names["var_q_moistening_ml"],
+        names["var_q_heating_ml"],
+        names["coord_z_center"],
+    )
+
+    # TODO Do all data merginig and loading before computing anything
+    init_times = list(set(ds_test[names["init_time_dim"]].values))
+    ds_hires = load_high_res_diag_dataset(
+        high_res_data_path,
+        init_times,
+        names["init_time_dim"],
+        names["renamed_hires_grid_vars"],
+    )
+    grid_path = os.path.join(os.path.dirname(test_data_path), "grid_spec.zarr")
+
+    # TODO ditto: do all merging of data before computing anything
+    grid = xr.open_zarr(fs_input.get_mapper(grid_path))
+    slmsk = ds_test[names["var_land_sea_mask"]].isel({names["init_time_dim"]: 0})
+
+    # TODO ditto: do all merging of data before computing anything
+    return merge_comparison_datasets(
+        data_vars=names["data_vars"],
+        datasets=[ds_pred, ds_test, ds_hires],
+        dataset_labels=[
+            DATASET_NAME_PREDICTION,
+            DATASET_NAME_FV3_TARGET,
+            DATASET_NAME_SHIELD_HIRES,
+        ],
+        grid=grid,
+        additional_dataset=slmsk,
+    )
 
 
 if __name__ == "__main__":
@@ -59,6 +148,9 @@ if __name__ == "__main__":
         "high_res_data_path",
         type=str,
         help="Path to C48 coarsened high res diagnostic data.",
+    )
+    parser.add_argument(
+        "variable_names_file", type=str, help="yml with variable name information"
     )
     parser.add_argument(
         "output_path",
@@ -79,13 +171,6 @@ if __name__ == "__main__":
         "The only type implemented right now is 'rf'.",
     )
     parser.add_argument(
-        "--delete-local-results-after-upload",
-        type=bool,
-        default=False,
-        help="If uploading to a remote results dir, delete the local copies"
-        " after upload.",
-    )
-    parser.add_argument(
         "--downsample-time-factor",
         type=int,
         default=1,
@@ -94,64 +179,34 @@ if __name__ == "__main__":
     args = parser.parse_args()
     args.test_data_path = os.path.join(args.test_data_path, "test")
 
-    # if output path is remote GCS location, save results to local output dir first
-    proto = get_protocol(args.output_path)
-    if proto == "" or proto == "file":
-        output_dir = args.output_path
-    elif proto == "gs":
-        remote_data_path, output_dir = os.path.split(args.output_path.strip("/"))
-    if not os.path.exists(output_dir):
-        os.mkdir(output_dir)
-    ds_test, ds_pred = predict_on_test_data(
+    with open(args.variable_names_file, "r") as f:
+        names = yaml.safe_load(f)
+
+    # remove trailing slash since in GCS some/path// is different than some/path/
+    output_dir = args.output_path.rstrip("/")
+
+    # TODO Lot's of input arguments and "and" in the name
+    # (potential target for refactor) At least it is clear
+    # that this can be improved now.
+    # Ideally, ML prediction and loading should be refactored to a separate routine
+    ds = load_data_and_predict_with_ml(
         args.test_data_path,
         args.model_path,
+        args.high_res_data_path,
         args.num_test_zarrs,
         args.model_type,
         args.downsample_time_factor,
+        names,
     )
 
-    fs_input = get_fs(args.test_data_path)
-    fs_output = get_fs(args.output_path)
+    # force loading now to avoid I/O issues down the line
+    # This could lead to OOM errors (but those sound like an issue anyway)
+    ds = ds.load()
 
-    add_column_heating_moistening(ds_test)
-    add_column_heating_moistening(ds_pred)
-    init_times = list(set(ds_test[INIT_TIME_DIM].values))
-    ds_hires = load_high_res_diag_dataset(args.high_res_data_path, init_times)
-    grid_path = os.path.join(os.path.dirname(args.test_data_path), "grid_spec.zarr")
-
-    grid = xr.open_zarr(fs_input.get_mapper(grid_path))
-    slmsk = ds_test["slmsk"].isel({INIT_TIME_DIM: 0})
-
-    ds = merge_comparison_datasets(
-        data_vars=DATA_VARS,
-        datasets=[ds_pred, ds_test, ds_hires],
-        dataset_labels=[
-            DATASET_NAME_PREDICTION,
-            DATASET_NAME_FV3_TARGET,
-            DATASET_NAME_SHIELD_HIRES,
-        ],
-        grid=grid,
-        additional_dataset=slmsk,
-    )
-    # separate datasets will now have common grid/sfc_type variables and
-    # an identifying dataset coordinate
-    ds_pred = ds.sel(dataset=DATASET_NAME_PREDICTION)
-    ds_test = ds.sel(dataset=DATASET_NAME_FV3_TARGET)
-    ds_hires = ds.sel(dataset=DATASET_NAME_SHIELD_HIRES)
-
-    ds_metrics = create_metrics_dataset(ds_pred, ds_test, ds_hires)
-    ds_metrics.to_netcdf(os.path.join(output_dir, "metrics.nc"))
-    metrics_plot_sections = plot_metrics(ds_metrics, output_dir, DPI_FIGURES)
-
-    diag_report_sections = plot_diagnostics(
-        ds_pred, ds_test, ds_hires, output_dir=output_dir, dpi_figures=DPI_FIGURES
-    )
-
-    combined_report_sections = {**metrics_plot_sections, **diag_report_sections}
-    create_report(combined_report_sections, "ml_offline_diagnostics", output_dir)
-
-    fs_output = get_fs(args.output_path)
-    if proto == "gs":
-        copy(output_dir, args.output_path)
-        if args.delete_local_results_after_upload is True:
-            shutil.rmtree(output_dir)
+    if _is_remote(args.output_path):
+        with tempfile.TemporaryDirectory() as local_dir:
+            # TODO another "and" indicates this needs to be refactored.
+            compute_metrics_and_plot(ds, local_dir, names)
+            copy(local_dir, output_dir)
+    else:
+        compute_metrics_and_plot(ds, args.output_path, names)