+++ HISTORY.rst
@@ -10,13 +10,23 @@ latest
 * Fixed integration tests to use same version tags of the `fv3net` and `prognostic_run` images
 * Added makefile targets to submit integration tests to cluster from local machine and to get docker image names
 * Made simple step output directory names the default in the orchestrator
+* Add `run_with_learned_nudging` workflow
+* Update fv3config submodule to v0.3.1
+* Add `get_config()` function to fv3net.runtime
+* Change API of `diagnostics_to_zarr` workflow so that it saves output zarrs in the given run directory
+* Add `nudge_to_obs` module to `kube_jobs`, which helps with the configuration of FV3GFS model runs that are nudged towards GFS analysis
 * Add public function: vcm.convert_timestamps
 * Add pipeline to load C384 restart data into a zarr
+* One step run workflow outputs a single zarr as output (instead of individual directories for each timestep), downstream workflows are adjusted to use this data format
+* Train data pipeline and offline diagnostics workflow read in variable names information from yaml provided to python
+* Force load data in diagnostics workflow before compute and plot
+* Improved logging when running FV3 model
+
 
 0.1.1 (2020-03-25)
 ------------------
 * Updates to make end-to-end workflow work with fv3atm (fv3gfs-python:v0.3.1)
-* Added bump2version for automated versioning of `fv3net` resources 
+* Added bump2version for automated versioning of `fv3net` resources
 * Add CircleCI build/push capabilities for `fv3net` images